Refactor common logic across new project and renewal processing into module
